# Adjust aggregations max_groups autosizing const