<!-- Thanks for sending a pull request!

If this is your first contribution, please review and sign our contributor agreement -
https://www.elastic.co/contributor-agreement.

Guidelines:
 - Prefer small PRs, and split changes into multiple logical commits where they must
   be delivered in a single PR.
 - If the PR is incomplete and not yet ready for review, open it as a Draft.
 - Once the PR is marked ready for review it is expected to pass all tests and linting,
   and you should not force-push any changes.

See also https://github.com/elastic/apm-server/blob/main/CONTRIBUTING.md for more tips on contributing.
-->

## Motivation/summary

<!--
Describe your change in the title and description, and provide a motivation for the
change and rationale for the approach taken.
-->

Adjust max_groups autosizing const such that the defaults are safe and will not lead to OOM.

According to the benchmarks, we will change the limits to:
- txmetrics max transaction groups: 10000 -> 5000 / 1GB memory limit
- txmetrics max service groups: 10000 -> 1000 / 1GB memory limit
- service tx metrics max groups: 10000 -> 1000 / 1GB memory limit
- service summary metrics max groups: 10000 -> 1000 / 1GB memory limit

---

OUTDATED (and all the benchmarks were based on):
Max groups autosizing will be adjusted as follows:
- txmetrics max transaction groups: 10000 -> 2500 / 1GB memory limit
- txmetrics max service groups: 10000 -> 500 / 1GB memory limit
- service tx metrics max groups: 10000 -> 500 / 1GB memory limit
- service summary metrics max groups: 10000 -> 500 / 1GB memory limit

### ~~Benchmark (on 8GB apm server, without histogram pooling)~~

**TLDR: max_rss 7857737728 -> 2369687552**

Before
```
Benchmark10000AggregationGroups-512          609        1151431598 ns/op                 0 error_responses/sec           0 errors/sec        13034 events/sec           87.00 gc_cycles        333.0 max_goroutines     7343182624 max_heap_alloc         37562153 max_heap_objects     7153254400 max_rss              16.09 mean_available_indexers          1525 metrics/sec        6391 spans/sec          5118 txs/sec      282429347 B/op   3339987 allocs/op
Benchmark10000AggregationGroups-512           84        1551815749 ns/op                 0 error_responses/sec           0 errors/sec        10411 events/sec            8.000 gc_cycles               121.0 max_goroutines     8013441408 max_heap_alloc         39208667 max_heap_objects     7670460416 max_rss              17.62 mean_available_indexers          1364 metrics/sec        4974 spans/sec          4074 txs/sec      296773386 B/op   3534176 allocs/op
Benchmark10000AggregationGroups-512           97        1525567700 ns/op                 0 error_responses/sec           0 errors/sec        10522 events/sec            9.000 gc_cycles               134.0 max_goroutines     7926363520 max_heap_alloc         41798615 max_heap_objects     7857737728 max_rss              17.56 mean_available_indexers          1167 metrics/sec        5094 spans/sec          4261 txs/sec      295408565 B/op   3545122 allocs/op
```

After
```
Benchmark10000AggregationGroups-512          127        1272911331 ns/op                 0 error_responses/sec           0 errors/sec        11925 events/sec           47.00 gc_cycles        146.0 max_goroutines     1885460424 max_heap_alloc         11475301 max_heap_objects     1970073600 max_rss              17.15 mean_available_indexers          1084 metrics/sec        6028 spans/sec          4813 txs/sec      274003331 B/op   3426760 allocs/op
Benchmark10000AggregationGroups-512           88        1477813998 ns/op                 0 error_responses/sec           0 errors/sec        10709 events/sec           32.00 gc_cycles        127.0 max_goroutines     2203089152 max_heap_alloc         13345200 max_heap_objects     2241953792 max_rss              17.46 mean_available_indexers          1048 metrics/sec        5255 spans/sec          4407 txs/sec      286051903 B/op   3523480 allocs/op
Benchmark10000AggregationGroups-512          114        1354900732 ns/op                 0 error_responses/sec           0 errors/sec        11388 events/sec           37.00 gc_cycles        139.0 max_goroutines     2261529672 max_heap_alloc         13466837 max_heap_objects     2369687552 max_rss              17.18 mean_available_indexers          1051 metrics/sec        5698 spans/sec          4640 txs/sec      287317164 B/op   3537602 allocs/op

```
---


### Updated benchmarks (with histogram pooling)

**1GB Summary: OOM -> 657711104**

1GB Before
OOM

1GB After
```
Benchmark10000AggregationGroups-64           106        3861961317 ns/op                 0 error_responses/sec           0 errors/sec         3309 events/sec          151.0 gc_cycles         131.0 max_goroutines      585402040 max_heap_alloc          3948170 max_heap_objects      657711104 max_rss               8.554 mean_available_indexers          461.5 metrics/sec              1829 spans/sec          1018 txs/sec      246313188 B/op   3048479 allocs/op
Benchmark10000AggregationGroups-64            25        5436423332 ns/op                 0 error_responses/sec           0 errors/sec         2375 events/sec           37.00 gc_cycles         79.00 max_goroutines     499902344 max_heap_alloc          3669938 max_heap_objects      587833344 max_rss               9.208 mean_available_indexers          313.1 metrics/sec              1308 spans/sec           754.4 txs/sec    250024680 B/op   3094211 allocs/op
Benchmark10000AggregationGroups-64            24        5416985959 ns/op                 0 error_responses/sec           0 errors/sec         2349 events/sec           35.00 gc_cycles         76.00 max_goroutines     505234048 max_heap_alloc          3807749 max_heap_objects      584384512 max_rss               9.222 mean_available_indexers          311.9 metrics/sec              1309 spans/sec           728.6 txs/sec    247219261 B/op   3060694 allocs/op
```

---

**2GB Summary: OOM -> 919334912**

2GB Before
OOM

2GB After
```
Benchmark10000AggregationGroups-128           85        1468469417 ns/op                 0 error_responses/sec           0 errors/sec         8726 events/sec           83.00 gc_cycles        117.0 max_goroutines      662553400 max_heap_alloc          4597885 max_heap_objects      758243328 max_rss               9.445 mean_available_indexers          865.2 metrics/sec              4844 spans/sec          3016 txs/sec      244917502 B/op   3069271 allocs/op
Benchmark10000AggregationGroups-128          126        1994455973 ns/op                 0 error_responses/sec           0 errors/sec         6509 events/sec          111.0 gc_cycles         133.0 max_goroutines      795960000 max_heap_alloc          5507223 max_heap_objects      919334912 max_rss               9.372 mean_available_indexers          709.4 metrics/sec              3567 spans/sec          2233 txs/sec      247377693 B/op   3085060 allocs/op
Benchmark10000AggregationGroups-128           42        3237547799 ns/op                 0 error_responses/sec           0 errors/sec         3968 events/sec           35.00 gc_cycles         88.00 max_goroutines     824741280 max_heap_alloc          5555509 max_heap_objects      901246976 max_rss              10.81 mean_available_indexers           419.7 metrics/sec              2191 spans/sec          1357 txs/sec      245981024 B/op   3065284 allocs/op
```

---

**4GB Summary: 4276580352 -> 1367863296**

4GB Before
```
Benchmark10000AggregationGroups-256          122        1088167937 ns/op                 0 error_responses/sec           0 errors/sec        12970 events/sec           26.00 gc_cycles        141.0 max_goroutines     3338803160 max_heap_alloc         16854174 max_heap_objects     3121729536 max_rss              10.86 mean_available_indexers          1281 metrics/sec        6649 spans/sec          5040 txs/sec      257424007 B/op   3191548 allocs/op
Benchmark10000AggregationGroups-256           79        1989790381 ns/op                 0 error_responses/sec           0 errors/sec         7527 events/sec           14.00 gc_cycles        114.0 max_goroutines     4385325704 max_heap_alloc         24268298 max_heap_objects     4276580352 max_rss              12.70 mean_available_indexers          1067 metrics/sec        3634 spans/sec          2825 txs/sec      275716575 B/op   3289389 allocs/op
Benchmark10000AggregationGroups-256          163        1347751663 ns/op                 0 error_responses/sec           0 errors/sec         4701 events/sec           33.00 gc_cycles        152.0 max_goroutines     3242247064 max_heap_alloc         16616866 max_heap_objects     2611253248 max_rss               8.051 mean_available_indexers         1179 metrics/sec        2039 spans/sec          1483 txs/sec      122145736 B/op   1327127 allocs/op
```

4GB After
```
Benchmark10000AggregationGroups-256          128        1242663147 ns/op                 0 error_responses/sec           0 errors/sec        11013 events/sec           78.00 gc_cycles        141.0 max_goroutines     1136689296 max_heap_alloc          7334993 max_heap_objects     1231147008 max_rss              12.13 mean_available_indexers           925.0 metrics/sec              5803 spans/sec          4285 txs/sec      250530933 B/op   3156959 allocs/op
Benchmark10000AggregationGroups-256           60        2452797301 ns/op                 0 error_responses/sec           0 errors/sec         5683 events/sec           36.00 gc_cycles        101.0 max_goroutines     1129519448 max_heap_alloc          7517340 max_heap_objects     1222520832 max_rss              13.25 mean_available_indexers           510.4 metrics/sec              2951 spans/sec          2222 txs/sec      261102855 B/op   3244121 allocs/op
Benchmark10000AggregationGroups-256           56        2414281263 ns/op                 0 error_responses/sec           0 errors/sec         5852 events/sec           30.00 gc_cycles        100.0 max_goroutines     1275729880 max_heap_alloc          8172653 max_heap_objects     1367863296 max_rss              13.20 mean_available_indexers           533.3 metrics/sec              2989 spans/sec          2330 txs/sec      262766475 B/op   3259861 allocs/op
```

## Checklist

<!--
Delete irrelevant items. The changelog should only be updated for user-facing changes.
Once the PR is ready for review there should be no unticked boxes.
-->

- [x] Update [CHANGELOG.asciidoc](https://github.com/elastic/apm-server/blob/main/CHANGELOG.asciidoc)
~~- [ ] Update [package changelog.yml](https://github.com/elastic/apm-server/blob/main/apmpackage/apm/changelog.yml) (only if changes to `apmpackage` have been made)~~
~~- [ ] Documentation has been updated~~

## How to test these changes

<!--
Explain how this PR can be tested by the reviewer: commands, dependencies, steps, etc.
If it is self-explanatory, delete this section.
-->

N/A


## Related issues

<!--
Reference the related issue(s), and make use of magic keywords where it makes sense
https://help.github.com/articles/closing-issues-using-keywords/.
-->
Part of #10147